requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Faker==17.6.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
validate-docbr==1.10.0
wrapt==1.12.1
